requirements.txt
streamlit
tensorflow
numpy
pandas
opencv-python
pillow
scikit-learn
matplotlib
kagglehub